Increased efficiency of update method
This makes changes to use the upsert attribute in mongodb so that
documents are only created if they do not already exist.

This change removes the need to call the find method to check if
a statement already exists in the database. This reduces the
number of database transactions that are required to complete
the operation.