This chatterbot build was built by Tomas Nelson<smallbusinessnetworks.tn@gmail.com
This chatterbot was built using pip, conda, npm, make, webpackstarter along with all dependencies necessary for the build.

This chatterbot is deemed the sole-property of O.S.M.Corp and use of it means you have paid for the version.
If you have not paid for the chatterbot(Hackbot) license then it is illegal to run this entity.  An entity being a artificial intelligent bot capable of search and destroy commands of zombies, hackbots, info gather bots, assist and decease hackers systems(databases,computer files, servers containing hacked files) etc.
If you paid then your use of it is strictly confidential and any use or changing of files will breach the contract and imply you used the project against its means and for a purpose of corruption in which the law will catch up with you.

If you have any questions or coments just know they are unnecessary and if your into sick things that your now the target of my build.